% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAViaR.R
\name{CAViaR}
\alias{CAViaR}
\title{CAViaR model estimation}
\usage{
CAViaR(
  Y,
  p = 1,
  q = 1,
  model.type = "SAV",
  tau = 0.05,
  band.hs = FALSE,
  jac.method = "Richardson",
  jac.options = list(),
  sign.level = 0.05,
  quant.type = 7,
  optim.config = list(),
  refine.estim = FALSE,
  refinement.conf = list()
)
}
\arguments{
\item{Y}{A vector, matrix, zoo or xts object containing the univariate series.}

\item{p}{Order of autoregressive quantile.}

\item{q}{Order of lag Y values.}

\item{model.type}{The specification, one of the following:
\itemize{
\item "SAV" (\emph{Symmetric absolute value})
}

\deqn{f_t(\theta) = \beta_0 + \sum_{i=1}^p \beta_i f_{t-i}(\beta) + \sum_{j=1}^q \gamma_j |y_{t-j}|}
\itemize{
\item "AS" (\emph{Asymmetric slope})
}

\deqn{f_t(\theta) = \beta_0 + \sum_{i=1}^p \beta_i f_{t-i}(\beta) + \sum_{j=1}^q \left( \gamma_{1,j} (y_{t-j})^+ + \gamma_{2,j} (y_{t-j})^- \right)}

where \eqn{(y_{t-j})^+ = \max(y_{t-j}, 0)} and \eqn{(y_{t-j})^- = \min(y_{t-j}, 0)}
\itemize{
\item "INDGARCH" (\emph{Indirect GARCH})
}

\deqn{f_t(\theta) = \left( \beta_0 + \sum_{i=1}^p \beta_i f_{t-i}^2(\beta) + \sum_{j=1}^q \gamma_j y_{t-j}^2 \right)^{1/2}}
\itemize{
\item "I-CAV" (\emph{Improved CAViaR})
}

\deqn{f_t(\theta) = \beta_0 + \beta_1 f_{t-1}(\theta) + (1 - \beta_1)\left(\frac{\nu}{1-\gamma_1} I(y_{t-1} > 0)+ \frac{\nu}{\gamma_1} I(y_{t-1} \leq 0)\right) |y_{t-1} - u|}

where \eqn{\nu = \sqrt{\gamma_1^2 + (1 - \gamma_1)^2}}, \eqn{0<\gamma_1<1} and \eqn{u} is the sample mean.}

\item{tau}{The quantile of interest. Set default to 0.05.}

\item{band.hs}{Logical parameter passed to \link[quantreg:bandwidth.rq]{quantreg::bandwidth.rq}, if \link{TRUE} the Hall-Sheather bandwidth is computed, if \link{FALSE} it is computed the Bofinger bandwidth.}

\item{jac.method}{One of "Richardson" (default), "simple", or "complex". This determines the method to compute finite differences jacobian to as part of the standard errors calculations. See \link[numDeriv:jacobian]{numDeriv::jacobian} for more information.}

\item{jac.options}{A list passed to \code{method.args} in \link[numDeriv:jacobian]{numDeriv::jacobian}.}

\item{sign.level}{The \code{alpha} parameter of \link[quantreg:bandwidth.rq]{quantreg::bandwidth.rq} function to control the level of significance for intended confidence intervals.}

\item{quant.type}{One of the types available in \link{quantile}. This is used to initialize the quantile process.}

\item{optim.config}{A list containing the optimization-related parameters, see \code{\link[nloptr:nloptr.print.options]{nloptr.print.options()}} for more information. See \code{details} for more information.}

\item{refine.estim}{Should the result from the global optimization be refined with a gradient-based solver?}

\item{refinement.conf}{A list containing the refinement optimization-related parameters, see \code{\link[nloptr:nloptr.print.options]{nloptr.print.options()}} for more information}
}
\value{
A list containing different results from the estimation. Class \code{CAViaR_estim}.
}
\description{
This function allows the estimation for a general order the caviar model proposed by Engle & Manganelli (2004).
}
\details{
The main difference in this implementation is the optimization procedure used to estimate the parameters. The original paper begins with a grid search and alternates between the Nelder-Mead and quasi-Newton optimization algorithms until convergence. In this function, however, a global optimization algorithm is used: the Multilevel Single Linkage ("NLOPT_GN_MLSL_LDS"). At each iteration, the sampled points are quasi-deterministic, using Sobol's low discrepancy sequences. Thus, different optimization runs are much less dependent on the choice of random seed.

As local optimizer, it is set by default to the \code{NELDER-MEAD}. However, other option is \code{SUBPLEX} which uses the \code{NELDER-MEAD} is subspaces, and is claimed to be more efficient and robust than the later. One can change this as described in the \link[nloptr:nloptr]{nloptr::nloptr} function.

An alternative to MLSL is the Improved Stochastic Ranking Evolution Strategy ("NLOPT_GN_ISRES"). However, this algorithm is much more dependent on the choice of the random seed. Although it is theoretically the case that it should return the same estimates among different runs, in practice, the user should be careful when choosing this algorithm. Furthermore, the standard errors are computed using the numerical Jacobian from \code{numDeriv}, and the bandwidth is computed as described in White et al. (2015).
}
\examples{
data=dataCAViaR
SAV <- CAViaR(Y=data$GM[1:2892],model.type = "SAV",
p=1,q=1,band.hs = TRUE,quant.type = 7,
tau=0.05,refine.estim = FALSE)

summary(SAV)
#or
SAV

plot(SAV)

}
\references{
Engle, R. F., & Manganelli, S. (2004). CAViaR: Conditional Autoregressive Value at Risk by Regression Quantiles. Journal of Business & Economic Statistics, 22(4), 367–381.

Huang, D., Yu, B., Fabozzi, F. J., & Fukushima, M. (2009). CAViaR-based forecast for oil price risk. Energy Economics, 31(4), 511-518.

White, H., Kim, T. H., & Manganelli, S. (2015). VAR for VaR: Measuring tail dependence using multivariate regression quantiles. Journal of econometrics, 187(1), 169-188.
}
