% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataCAViaR}
\alias{dataCAViaR}
\title{Original caviar data used in Engle & Manganelli (2004).}
\format{
3392 rows x 3 columns of log returns of General Motors, IBM and S&P 500.
}
\source{
\url{https://www.simonemanganelli.org/Simone/Research.html}
}
\usage{
dataCAViaR
}
\description{
Original caviar data used in Engle & Manganelli (2004).
}
\references{
Engle, R. F., & Manganelli, S. (2004). CAViaR: Conditional autoregressive value at risk by regression quantiles. Journal of business & economic statistics, 22(4), 367-381.
}
\keyword{datasets}
