% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MVMQ}
\alias{MVMQ}
\title{Data used in White et al. (2015)}
\format{
An xts object containing the returns of Goldman sachs, Braclays, HSBC and Deutsche Bank, as well as the corresponding index.
}
\source{
\url{https://www.simonemanganelli.org/Simone/Research.html}
}
\usage{
MVMQ
}
\description{
Data used in White et al. (2015)
}
\references{
White, H., Kim, T. H., & Manganelli, S. (2015). VAR for VaR: Measuring tail dependence using multivariate regression quantiles. Journal of econometrics, 187(1), 169-188.
}
\keyword{datasets}
