% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMQ-CAViaR.R
\name{MVMQ_CAViaR}
\alias{MVMQ_CAViaR}
\title{MVMQ-CAViaR model estimation proposed by White et al. (2015).}
\usage{
MVMQ_CAViaR(
  Y,
  p = 1,
  q = 1,
  tau = rep(0.05, ncol(Y)),
  band.hs = FALSE,
  jac.method = "simple",
  jac.options = list(),
  sign.level = 0.05,
  quant.type = 7,
  optim.config = list(),
  init.optim.config = list(),
  global.estim = FALSE,
  global.optim = list()
)
}
\arguments{
\item{Y}{A matrix, xts or zoo object of multivariate series.}

\item{p}{Order of autoregressive quantile.}

\item{q}{Order of lag Y values.}

\item{tau}{A vector containing the quantiles of interest. Set default to 0.05.}

\item{band.hs}{Logical parameter passed to \link[quantreg:bandwidth.rq]{quantreg::bandwidth.rq}, if \link{TRUE} the Hall-Sheather bandwidth is computed, if \link{FALSE} it is computed the Bofinger bandwidth.}

\item{jac.method}{One of "Richardson" (default), "simple", or "complex". This determines the method to compute finite differences jacobian to as part of the standard errors calculations. See \link[numDeriv:jacobian]{numDeriv::jacobian} for more information.}

\item{jac.options}{A list pased to \code{method.args} in \link[numDeriv:jacobian]{numDeriv::jacobian}.}

\item{sign.level}{The \code{alpha} parameter of \link[quantreg:bandwidth.rq]{quantreg::bandwidth.rq} function to control the level of significance for intended confidence intervals.}

\item{quant.type}{One of the types available in \link{quantile}. This is used to initialize the quantile process.}

\item{optim.config}{A list containing the optimization-related parameters, see \code{\link[nloptr:nloptr.print.options]{nloptr.print.options()}} and \code{details} for more information.}

\item{init.optim.config}{A list containing the univariate optimization-related parameters, see \code{\link[nloptr:nloptr.print.options]{nloptr.print.options()}} for more information.}

\item{global.estim}{Should the optimization be permormed in first place with global optimization with \link[GenSA:GenSA]{GenSA::GenSA} in order to potentially achieve a better exploration of parameter space?}

\item{global.optim}{A list containing the global optimization-related parameters, see \link[GenSA:GenSA]{GenSA::GenSA}  for more information.}
}
\value{
A list containing different results from the estimation. Class \code{MVMQ_CAViaR}
}
\description{
MVMQ-CAViaR model estimation proposed by White et al. (2015).
}
\details{
This implementation follows in essence the same optimization strategy as White et al. (2015), by first obtaining for each series their univariate CAViaR estimation with \link{CAViaR}, and setting the rest of the parameters to 0. As mentioned above, staring from those univariate estimates, this function offers two different optimization strategies: i) feed this starting point to a local optimizer (by default \code{NLOPT_LN_SBPLX}) following the original work, in such case the user should set \code{global.estim}=FALSE (the default); ii) feed this starting point to \link[GenSA:GenSA]{GenSA::GenSA} in order perform a global optimization with the intention to explore the parameter space, and the result from this global phase is then used as starting point for the same procedure as i), then the user should set \code{global.estim}=TRUE.

Regarding the specification, at the moment, only the symmetric absolute value specification is avaliable, having the following form:
\deqn{\boldsymbol{f_t}(\boldsymbol{\theta}) = \boldsymbol{c} + \sum_{i=1}^p \boldsymbol{A_i} \boldsymbol{f_{t-i}(\boldsymbol{\theta})} + \sum_{j=1}^q \boldsymbol{B_j} |\boldsymbol{Y_{t-j}|}}
}
\examples{
Barclays <- MVMQ_CAViaR(MVMQ[,c(6,1)],tau =c(0.01,0.01),band.hs = TRUE)
summary(Barclays)
#or
Barclays
plot(Barclays,rows=2,columns=1)

}
\references{
White, H., Kim, T. H., & Manganelli, S. (2015). VAR for VaR: Measuring tail dependence using multivariate regression quantiles. Journal of econometrics, 187(1), 169-188.
}
