% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.CAViaR_estim}
\alias{plot.CAViaR_estim}
\title{Plot CAViaR}
\usage{
\method{plot}{CAViaR_estim}(x, .by = "month", .format = "\%b-\%Y", titl = "VaR", ...)
}
\arguments{
\item{x}{\code{CAViaR_estim} class object.}

\item{.by}{Frequency to diplay the dates in the plot, see \link{axis.Date}.}

\item{.format}{Format of the displayed dates in the plot, see \link{axis.Date}.}

\item{titl}{Optional title of the plot.}

\item{...}{other arguments to plot}
}
\description{
Plot CAViaR
}
