% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.MVMQ_CAViaR}
\alias{plot.MVMQ_CAViaR}
\title{Plot MVMQ_CAViaR}
\usage{
\method{plot}{MVMQ_CAViaR}(
  x,
  rows = 2,
  columns = 1,
  .by = "month",
  .format = "\%b-\%Y",
  titl = "VaR at ",
  ...
)
}
\arguments{
\item{x}{\code{MVMQ_CAViaR} class object.}

\item{rows}{Number of rows to display in the plot, passed to \link{par}}

\item{columns}{Number of columns to display in the plot, passed to \link{par}}

\item{.by}{Frequency to diplay the dates in the plot, see \link{axis.Date}.}

\item{.format}{Format of the displayed dates in the plot, see \link{axis.Date}.}

\item{titl}{Optional title of the plot.}

\item{...}{other arguments to plot}
}
\description{
Plot MVMQ_CAViaR
}
