% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.CAViaR_estim}
\alias{print.CAViaR_estim}
\title{Print CAViaR}
\usage{
\method{print}{CAViaR_estim}(x, digits = 5, conf.level = 0.95, ...)
}
\arguments{
\item{x}{\code{CAViaR_estim} class object.}

\item{digits}{number of decimals to display, passed to \link{round}.}

\item{conf.level}{Confidence level to the coverage test.}

\item{...}{Other arguments passed to \link{print}.}
}
\description{
Print CAViaR
}
