% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.MVMQ_CAViaR}
\alias{print.MVMQ_CAViaR}
\title{Print MVMQ_CAViaR}
\usage{
\method{print}{MVMQ_CAViaR}(x, digits = 5, conf.level = 0.95, ...)
}
\arguments{
\item{x}{\code{MVMQ_CAViaR} class object.}

\item{digits}{number of decimals to display, passed to \link{round}.}

\item{conf.level}{Confidence level to the coverage test.}

\item{...}{Other arguments passed to \link{print}.}
}
\description{
Print MVMQ_CAViaR
}
