% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.CAViaR_estim}
\alias{summary.CAViaR_estim}
\title{Summary CAViaR}
\usage{
\method{summary}{CAViaR_estim}(object, digits = 5, conf.level = 0.95, ...)
}
\arguments{
\item{object}{\code{CAViaR_estim} class object.}

\item{digits}{number of decimals to display, passed to \link{round}.}

\item{conf.level}{Confidence level to the coverage test.}

\item{...}{Other arguments passed to \link{print}.}
}
\description{
Summary CAViaR
}
