% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.MVMQ_CAViaR}
\alias{summary.MVMQ_CAViaR}
\title{Summary MVMQ_CAViaR}
\usage{
\method{summary}{MVMQ_CAViaR}(object, digits = 5, conf.level = 0.95, ...)
}
\arguments{
\item{object}{\code{MVMQ_CAViaR} class object.}

\item{digits}{number of decimals to display ,passed to \link{round}.}

\item{conf.level}{Confidence level to the coverage test.}

\item{...}{Other arguments passed to \link{print}.}
}
\value{
No value returned
}
\description{
Summary MVMQ_CAViaR
}
