% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wti_data}
\alias{wti_data}
\title{Data used in Huang et al. (2009).}
\format{
An xts object of the WTI price daily returns
}
\source{
\url{https://fred.stlouisfed.org/series/WTISPLC}
}
\usage{
wti_data
}
\description{
Data used in Huang et al. (2009).
}
\references{
Huang, D., Yu, B., Fabozzi, F. J., & Fukushima, M. (2009). CAViaR-based forecast for oil price risk. Energy Economics, 31(4), 511-518.
}
\keyword{datasets}
